fix: make invoking the v8 platform initialize not dependent with cli
